requirements.txt
ccxt>=4.0.0
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
loguru>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.8.0
websockets>=11.0.0
typing-extensions>=4.0.0

# Web GUI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
starlette>=0.27.0
sse-starlette>=1.6.0
python-socketio>=5.9.0
aiofiles>=23.2.0

# Additional utilities for web
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
